+++ balance_car/Two_balance_car/Core/Src/gpio.c
@@ -53,8 +53,8 @@ void MX_GPIO_Init(void)
   HAL_GPIO_WritePin(GPIOA, SR04_TRIG_Pin|DRV_EN_Pin, GPIO_PIN_RESET);
 
   /*Configure GPIO pin Output Level */
-  HAL_GPIO_WritePin(GPIOB, BIN1_Pin|BIN2_Pin|AIN1_Pin|AIN2_Pin
-                          |OLED_SCL_Pin|OLED_SDA_Pin, GPIO_PIN_RESET);
+  HAL_GPIO_WritePin(GPIOB, IMU_SCL_Pin|IMU_SDA_Pin|BIN1_Pin|BIN2_Pin
+                          |AIN1_Pin|AIN2_Pin|OLED_SCL_Pin|OLED_SDA_Pin, GPIO_PIN_RESET);
 
   /*Configure GPIO pins : SR04_TRIG_Pin DRV_EN_Pin */
   GPIO_InitStruct.Pin = SR04_TRIG_Pin|DRV_EN_Pin;
@@ -75,10 +75,10 @@ void MX_GPIO_Init(void)
   GPIO_InitStruct.Pull = GPIO_PULLUP;
   HAL_GPIO_Init(ESTOP_GPIO_Port, &GPIO_InitStruct);
 
-  /*Configure GPIO pins : BIN1_Pin BIN2_Pin AIN1_Pin AIN2_Pin
-                           OLED_SCL_Pin OLED_SDA_Pin */
-  GPIO_InitStruct.Pin = BIN1_Pin|BIN2_Pin|AIN1_Pin|AIN2_Pin
-                          |OLED_SCL_Pin|OLED_SDA_Pin;
+  /*Configure GPIO pins : IMU_SCL_Pin IMU_SDA_Pin BIN1_Pin BIN2_Pin
+                           AIN1_Pin AIN2_Pin OLED_SCL_Pin OLED_SDA_Pin */
+  GPIO_InitStruct.Pin = IMU_SCL_Pin|IMU_SDA_Pin|BIN1_Pin|BIN2_Pin
+                          |AIN1_Pin|AIN2_Pin|OLED_SCL_Pin|OLED_SDA_Pin;
   GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
   GPIO_InitStruct.Pull = GPIO_NOPULL;
   GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_HIGH;